@@ -19,6 +19,8 @@
     "@angular/platform-browser": "~4.0.0",
     "@angular/platform-browser-dynamic": "~4.0.0",
     "@angular/router": "~4.0.0",
+    "@ngrx/effects": "^4.0.1",
+    "@ngrx/store": "^4.0.0",
     "email-validator": "^1.0.7",
     "nativescript-angular": "~3.0.0",
     "nativescript-appversion": "^1.4.1",
@@ -39,6 +41,6 @@
     "lazy": "1.0.11",
     "nativescript-dev-android-snapshot": "^0.*.*",
     "nativescript-dev-typescript": "^0.4.0",
-    "typescript": "~2.2.1"
+    "typescript": "~2.3.0"
   }
 }
