@@ -1,9 +1,10 @@
 import {Component, OnInit} from "@angular/core";
 import { Page } from "ui/page";
-import {Config} from "../../shared/config";
+import {Config} from "../../modules/core/config";
 import {User} from "../../shared/user/user.class";
 import {Router} from "@angular/router";
 import {UserService} from "../../shared/user/user.service";
+import {IUser} from "../../shared/user/user.model";
 
 
 @Component({
@@ -14,13 +15,16 @@ import {UserService} from "../../shared/user/user.service";
 })
 
 export class CardInfoComponent implements OnInit {
-    user: User;
+    user:IUser;
 
     constructor(private page: Page,
                 private router: Router,
                 private userService: UserService,){
 
         this.user = new User();
+
+
+        /* @TODO: create Requset -> get User Profile from server & subscribe */
     }
 
     goToSettings(){
@@ -34,13 +38,11 @@ export class CardInfoComponent implements OnInit {
     ngOnInit(){
         this.page.androidStatusBarBackground = Config.ActionBarColor;
 
-        this.userService.getProfile().subscribe(({ signal, name, balance, date } : User)=>{
-            this.user.signal = signal;
-            this.user.name = name;
-            this.user.balance = balance;
-            this.user.date = date;
-        });
-
-        /* @TODO: create Requset -> get User Profile from server & subscribe */
+        // this.userService.getProfile().subscribe(({ signal, name, balance, date } : User)=>{
+        //     this.user.signal = signal;
+        //     this.user.name = name;
+        //     this.user.balance = balance;
+        //     this.user.date = date;
+        // });
     }
 }
