@@ -1,6 +1,6 @@
-import {AutoparkInterface} from "./autopark.model";
+import {IAutopark} from "./autopark.model";
 
-export class Autopark implements  AutoparkInterface{
+export class Autopark implements  IAutopark{
     constructor(public city, public name, public title){
     }
 }
