@@ -1,4 +1,4 @@
-export interface AutoparkInterface{
+export interface IAutopark{
     city: string;
     name: string;
     title: string;
