@@ -3,7 +3,7 @@ import {Http, ConnectionBackend, RequestOptions, Request, RequestOptionsArgs, Re
 import {Observable} from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
-import {Config} from "./config";
+import {Config} from "../modules/core/config";
 
 @Injectable()
 export class InterceptedHttp extends Http {
@@ -47,7 +47,7 @@ export class InterceptedHttp extends Http {
             options.headers = new Headers();
         }
         options.headers.append('Content-Type', 'application/json');
-        options.headers.append('token', localStorage.getItem("token"));
+        // options.headers.append('X-Driver-Auth', localStorage.getItem("token"));
 
         return options;
     }
